+++ packages/mattermost-plugin/selectors.ts
@@ -1,7 +1,6 @@
+import type {GlobalState} from '@mattermost/types/store'
 import {getConfig} from 'mattermost-redux/selectors/entities/general'
 import {getCurrentTeam} from 'mattermost-redux/selectors/entities/teams'
-
-import type {GlobalState} from 'mattermost-redux/types/store'
 
 import manifest from './manifest'